Update sdpa_io.cpp
It doesn't compile under GCC 11 because of the line `LP_CNonZeroCount[blockNumber[l2]+j]++;`, as it is trying to increment a bool. I don't know what it is supposed to do, whether this variable should have been an int, or whether you just want to set it to true. In any case, since the function has a "not use" comment, perhaps it's better to remove obsolete code entirely.

Note that the same issue affects all other versions of SDPA, but I don't want to create the same pull request over and over again.